adicionando teste e preparando deploy

+++ src/main/java/com/generation/blogpessoal/controller/PostagemController.java
@@ -1,6 +1,7 @@
 package com.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,7 +14,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.model.Postagem;
 import com.generation.blogpessoal.repository.PostagemRepository;
@@ -25,68 +28,66 @@ import jakarta.validation.Valid;
 @RequestMapping("/postagens")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class PostagemController {
-
 	@Autowired
 	private PostagemRepository postagemRepository;
 	
 	@Autowired
 	private TemaRepository temaRepository;
-
+	
 	@GetMapping
-	public ResponseEntity<List<Postagem>> getAll() {
+	public ResponseEntity<List<Postagem>> getAll(){
+		// SELECT * FROM tb_postagens
 		return ResponseEntity.ok(postagemRepository.findAll());
-
-		// SELECT * FROM tb_postagens;
 	}
-
-	@GetMapping("/{id}")
-	public ResponseEntity<Postagem> getById(@PathVariable Long id) {
+	
+	@GetMapping("{id}")
+	public ResponseEntity<Postagem> getById(@PathVariable Long id){
+		// SELECT * FROM tb_postagens WHERE id = ?
 		return postagemRepository.findById(id)
 				.map(resposta -> ResponseEntity.ok(resposta))
 				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-
-		// SELECT * FROM tb_postagens WHERE id=?;
-
-	}
-
-	@GetMapping("/titulo/{titulo}")
-	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo) {
-
-		return ResponseEntity.ok(postagemRepository.findAllByTituloContainingIgnoreCase(titulo));
-
-		// SELECT * FROM tb_postagens WHERE titulo LIKE "%titulo%";
-	}
-
-	@PostMapping
-	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem) {
-	    if (!temaRepository.existsById(postagem.getTema().getId())) {
-	        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
-	    }
-	    return ResponseEntity.status(HttpStatus.CREATED).body(postagemRepository.save(postagem));
-	}
-
-	@PutMapping
-	public ResponseEntity<?> put(@Valid @RequestBody Postagem postagem){
-	    if (!temaRepository.existsById(postagem.getTema().getId())) {
-	        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
-	    }
-	    return postagemRepository.findById(postagem.getId())
-	            .map(resposta -> {
-	                postagemRepository.save(postagem);
-	                return ResponseEntity.status(HttpStatus.OK).build();
-	            })
-	            .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 	
-	@DeleteMapping("/{id}")
-	public ResponseEntity<?> deleteById(@PathVariable Long id) { 
-		return postagemRepository.findById(id)
-				.map(resposta -> {
-					postagemRepository.deleteById(id);
-					return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-				})
+	@GetMapping("/titulo/{titulo}")
+	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo){
+		// SELECT * FROM tb_postagens WHERE titulo LIKE "%???%"
+		return ResponseEntity.ok(postagemRepository.findAllByTituloContainingIgnoreCase(titulo));
+	}
+	
+	@PostMapping
+	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem){
+		// INSERT INTO tb_postagens(titulo, texto) VALUES ("???", "???")
+		
+		return temaRepository.findById(postagem.getTema().getId())
+				.map(resposta -> ResponseEntity.status(HttpStatus.CREATED).body(postagemRepository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
+	}
+	
+	@PutMapping
+	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem){
+		// UPDATE tb_postagens SET titulo = "???", texto = "???", data = "???" WHERE id = ?
+		
+		return postagemRepository.findById(postagem.getId())
+				.map(resposta -> 
+					temaRepository.findById(postagem.getTema().getId())
+						.map(resposta2 -> ResponseEntity.status(HttpStatus.OK).body(postagemRepository.save(postagem)))
+						.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build())
+				)
 				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-
-		// DELETE FROM tb_postagens WHERE id = id;
+	}
+	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("{id}")
+	public void delete(@PathVariable Long id) {
+		// DELETE FROM tb_postagens WHERE id = ?
+		
+		Optional<Postagem> optional = postagemRepository.findById(id);
+		
+		if (optional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		else {
+			postagemRepository.deleteById(id);
+		}
 	}
 }
